+++ packages/client/src/components/app/blocks/FormBlockComponent.svelte
@@ -16,7 +16,7 @@
     [FieldType.LONGFORM]: "longformfield",
     [FieldType.DATETIME]: "datetimefield",
     [FieldType.ATTACHMENT]: "attachmentfield",
-    [FieldType.ATTACHMENT_SINGLE]: "singleattachmentfield",
+    [FieldType.ATTACHMENT_SINGLE]: "attachmentsinglefield",
     [FieldType.LINK]: "relationshipfield",
     [FieldType.JSON]: "jsonfield",
     [FieldType.BARCODEQR]: "codescanner",
