+++ packages/client/src/components/app/forms/index.js
@@ -9,7 +9,7 @@ export { default as booleanfield } from "./BooleanField.svelte"
 export { default as longformfield } from "./LongFormField.svelte"
 export { default as datetimefield } from "./DateTimeField.svelte"
 export { default as attachmentfield } from "./AttachmentField.svelte"
-export { default as singleattachmentfield } from "./SingleAttachmentField.svelte"
+export { default as attachmentsinglefield } from "./AttachmentSingleField.svelte"
 export { default as relationshipfield } from "./RelationshipField.svelte"
 export { default as passwordfield } from "./PasswordField.svelte"
 export { default as formstep } from "./FormStep.svelte"
